feat: CouponIssuePolicy 추가

+++ src/main/java/com/teemo/shopping/coupon/service/CouponService.java
@@ -1,13 +1,22 @@
 package com.teemo.shopping.coupon.service;
 
+import com.teemo.shopping.account.domain.Account;
+import com.teemo.shopping.account.domain.AccountsCoupons;
+import com.teemo.shopping.account.repository.AccountRepository;
+import com.teemo.shopping.account.repository.AccountsCouponsRepository;
 import com.teemo.shopping.coupon.domain.Coupon;
+import com.teemo.shopping.coupon.domain.CouponIssuePoliciesAccounts;
+import com.teemo.shopping.coupon.domain.CouponIssuePolicy;
 import com.teemo.shopping.coupon.dto.CouponDTO;
+import com.teemo.shopping.coupon.repository.CouponIssuePoliciesAccountsRepository;
+import com.teemo.shopping.coupon.repository.CouponIssuePolicyRepository;
 import com.teemo.shopping.coupon.repository.CouponRepository;
-import com.teemo.shopping.game.domain.Game;
-import com.teemo.shopping.game.dto.GameDTO;
-import com.teemo.shopping.game.repository.GameRepository;
+import java.time.LocalDateTime;
+import java.util.Date;
 import java.util.List;
+import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.parameters.P;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -15,9 +24,8 @@ import org.springframework.transaction.annotation.Transactional;
 public class CouponService {
     @Autowired
     private CouponRepository couponRepository;
-
     @Transactional
-    public Long add(CouponDTO couponDTO) throws Exception {
+    public Long add(CouponDTO couponDTO) throws Exception { //todo: game, gameCategory 추가 하기
         Coupon coupon = couponDTO.to();
         coupon = couponRepository.save(coupon);
         return coupon.getId();
@@ -25,6 +33,9 @@ public class CouponService {
 
     @Transactional
     public void remove(Long couponId) throws Exception {
+        if(!couponRepository.existsById(couponId)) {
+            throw new IllegalStateException("존재하지 않는 쿠폰임.");
+        }
         couponRepository.deleteById(couponId);
     }
